Move naga DSL tests into sibling test module
