+++ src/compiler/wasm/rust/oscilla-naga-shim/src/lowering/mod.rs
@@ -2,3 +2,6 @@ pub mod dsl;
 
 #[allow(unused_imports)]
 pub use dsl::{Expr, FnBodyBuilder, FnBuilder, ModuleBuilder};
+
+#[cfg(test)]
+mod dsl_tests;
